+++ src/utils/gameParameters.ts
